envoy: Optimize list of allowed remote security IDs
We omit explicit listing of allowed remote security IDs when they are
for a wildcard selector. Do the same if there is only one selector for
the port, as bpf l3/l4 will pass only allowed l3s through. If there
are multiple selectors, then we need to differentiate between remotes
based on the L7 rules, so we must explicitly list the allowed remotes
for non-wildcard selectors in that case.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>