api/peer: add TLS field to ChangeNotification message
Add a new field to `ChangeNotification` to provide information to
establish a TLS connection to the peer. If unset, the connection should
be assumed to only be available as plaintext.

This change is backward compatible. For the time being, only the server
name to match against the certificate is required. More information may
be added to the new TLS message later if needed.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>